+++ apps/probability/law/normal_law.cpp
@@ -75,7 +75,7 @@ float NormalLaw::evaluateAtAbscissa(float x) const {
   return (1.0f/(std::fabs(m_parameter2)*std::sqrt(2.0f*M_PI)))*std::exp(-0.5f*std::pow((x-m_parameter1)/m_parameter2,2));
 }
 
-bool NormalLaw::authorizedValueAtIndex(double x, int index) const {
+bool NormalLaw::authorizedValueAtIndex(float x, int index) const {
   if (index == 0) {
     return true;
   }
